+++ src/backend/tests/partaj/core/test_api_referral_request_answer_validation.py
@@ -1,0 +1,463 @@
+from unittest import mock
+import uuid
+
+from django.conf import settings
+from django.test import TestCase
+
+from rest_framework.authtoken.models import Token
+
+from partaj.core import factories, models
+
+
+@mock.patch("partaj.core.email.Mailer.send")
+class ReferralApiRequestAnswerValidationTestCase(TestCase):
+    """
+    Test API routes and actions related to the Referral request answer validation endpoint.
+    """
+
+    def test_referral_request_answer_validation_by_anonymous_user(
+        self, mock_mailer_send
+    ):
+        """
+        Anonymous users cannot request a validation on an answer for a referral.
+        """
+        referral = factories.ReferralFactory(state=models.ReferralState.PROCESSING)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+        )
+        self.assertEqual(response.status_code, 401)
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.PROCESSING)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_by_random_logged_in_user(
+        self, mock_mailer_send
+    ):
+        """
+        Any random logged in user cannot request a validation on an answer for a referral.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.PROCESSING)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+        self.assertEqual(response.status_code, 403)
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.PROCESSING)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_by_linked_user(self, mock_mailer_send):
+        """
+        The linked user cannot request a validation on an answer for a referral.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(
+            state=models.ReferralState.PROCESSING, post__users=[user]
+        )
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 403)
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.PROCESSING)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_by_linked_unit_member(
+        self, mock_mailer_send
+    ):
+        """
+        Linked unit members can request a validation on an answer for a referral.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.PROCESSING)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json()["state"], models.ReferralState.IN_VALIDATION)
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 1)
+        # Make sure the validation request was built with the data we expect
+        validation_request = models.ReferralAnswerValidationRequest.objects.get(
+            answer=answer,
+            validator=validator,
+        )
+        # An activity was created for this validation request
+        self.assertEqual(
+            models.ReferralActivity.objects.get(
+                verb=models.ReferralActivityVerb.VALIDATION_REQUESTED
+            ).item_content_object.id,
+            validation_request.id,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.IN_VALIDATION)
+        mock_mailer_send.assert_called_with(
+            {
+                "params": {
+                    "case_number": referral.id,
+                    "created_by": user.get_full_name(),
+                    "link_to_referral": (
+                        f"https://partaj/app/unit/{referral.units.get().id}"
+                        f"/referrals-list/referral-detail/{referral.id}"
+                    ),
+                    "referral_users": referral.users.first().get_full_name(),
+                    "title": referral.object,
+                    "topic": referral.topic.name,
+                    "unit_name": referral.units.get().name,
+                },
+                "replyTo": {"email": "contact@partaj.beta.gouv.fr", "name": "Partaj"},
+                "templateId": settings.SENDINBLUE[
+                    "REFERRAL_ANSWER_VALIDATION_REQUESTED_TEMPLATE_ID"
+                ],
+                "to": [{"email": validator.email}],
+            }
+        )
+
+    def test_referral_request_duplicate_answer_validation(self, mock_mailer_send):
+        """
+        An error should be raised if a user attempts to request a validation for an answer from
+        a user who was already requested one.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.IN_VALIDATION)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory(first_name="Alfred", last_name="Borden")
+
+        # Create an existing validation request for the same answer and validator
+        factories.ReferralAnswerValidationRequestFactory(
+            answer=answer, validator=validator
+        )
+        self.assertEqual(
+            models.ReferralAnswerValidationRequest.objects.all().count(), 1
+        )
+
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {"errors": ["Alfred Borden was already requested to validate this answer"]},
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 1)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.IN_VALIDATION)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_nonexistent_answer_validation_by_linked_unit_member(
+        self, mock_mailer_send
+    ):
+        """
+        An explicit error is raised when a unit member attempts to request a validation for an
+        answer that does not exist.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.IN_VALIDATION)
+        referral.units.get().members.add(user)
+        random_uuid = uuid.uuid4()
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": random_uuid, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json()["errors"], [f"answer {random_uuid} does not exist"]
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.IN_VALIDATION)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_by_linked_unit_member_from_nonexistent_user(
+        self, mock_mailer_send
+    ):
+        """
+        An explicit error is raised when a unit member attempts to request a validation from a
+        user that does not exist.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.PROCESSING)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        random_uuid = uuid.uuid4()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": random_uuid},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json()["errors"], [f"user {random_uuid} does not exist"]
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(models.ReferralActivity.objects.count(), 0)
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.PROCESSING)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_from_in_validation_state(
+        self, mock_mailer_send
+    ):
+        """
+        New answer validations can be requested for a referral already in the
+        IN_VALIDATION state.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.IN_VALIDATION)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json()["state"], models.ReferralState.IN_VALIDATION)
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 1)
+        # Make sure the validation request was built with the data we expect
+        validation_request = models.ReferralAnswerValidationRequest.objects.get(
+            answer=answer,
+            validator=validator,
+        )
+        # An activity was created for this validation request
+        self.assertEqual(
+            models.ReferralActivity.objects.get(
+                verb=models.ReferralActivityVerb.VALIDATION_REQUESTED
+            ).item_content_object.id,
+            validation_request.id,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.IN_VALIDATION)
+        mock_mailer_send.assert_called_with(
+            {
+                "params": {
+                    "case_number": referral.id,
+                    "created_by": user.get_full_name(),
+                    "link_to_referral": (
+                        f"https://partaj/app/unit/{referral.units.get().id}"
+                        f"/referrals-list/referral-detail/{referral.id}"
+                    ),
+                    "referral_users": referral.users.first().get_full_name(),
+                    "title": referral.object,
+                    "topic": referral.topic.name,
+                    "unit_name": referral.units.get().name,
+                },
+                "replyTo": {"email": "contact@partaj.beta.gouv.fr", "name": "Partaj"},
+                "templateId": settings.SENDINBLUE[
+                    "REFERRAL_ANSWER_VALIDATION_REQUESTED_TEMPLATE_ID"
+                ],
+                "to": [{"email": validator.email}],
+            }
+        )
+
+    def test_referral_request_answer_validation_from_received_state(
+        self, mock_mailer_send
+    ):
+        """
+        New answer validations cannot be requested for a referral in the
+        RECEIVED state.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.RECEIVED)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {
+                "errors": [
+                    "Transition REQUEST_ANSWER_VALIDATION not allowed from state received."
+                ]
+            },
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(
+            models.ReferralActivity.objects.count(),
+            0,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.RECEIVED)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_from_assigned_state(
+        self, mock_mailer_send
+    ):
+        """
+        New answer validations cannot be requested for a referral in the
+        ASSIGNED state.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.ASSIGNED)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {
+                "errors": [
+                    "Transition REQUEST_ANSWER_VALIDATION not allowed from state assigned."
+                ]
+            },
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(
+            models.ReferralActivity.objects.count(),
+            0,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.ASSIGNED)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_from_answered_state(
+        self, mock_mailer_send
+    ):
+        """
+        New answer validations cannot be requested for a referral in the
+        ANSWERED state.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.ANSWERED)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {
+                "errors": [
+                    "Transition REQUEST_ANSWER_VALIDATION not allowed from state answered."
+                ]
+            },
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(
+            models.ReferralActivity.objects.count(),
+            0,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.ANSWERED)
+        mock_mailer_send.assert_not_called()
+
+    def test_referral_request_answer_validation_from_closed_state(
+        self, mock_mailer_send
+    ):
+        """
+        New answer validations cannot be requested for a referral in the
+        CLOSED state.
+        """
+        user = factories.UserFactory()
+        referral = factories.ReferralFactory(state=models.ReferralState.CLOSED)
+        answer = factories.ReferralAnswerFactory(
+            referral=referral,
+            state=models.ReferralAnswerState.DRAFT,
+        )
+        referral.units.get().members.add(user)
+        validator = factories.UserFactory()
+        response = self.client.post(
+            f"/api/referrals/{referral.id}/request_answer_validation/",
+            {"answer": answer.id, "validator": validator.id},
+            HTTP_AUTHORIZATION=f"Token {Token.objects.get_or_create(user=user)[0]}",
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {
+                "errors": [
+                    "Transition REQUEST_ANSWER_VALIDATION not allowed from state closed."
+                ]
+            },
+        )
+        self.assertEqual(models.ReferralAnswerValidationRequest.objects.count(), 0)
+        self.assertEqual(
+            models.ReferralActivity.objects.count(),
+            0,
+        )
+        referral.refresh_from_db()
+        self.assertEqual(referral.state, models.ReferralState.CLOSED)
+        mock_mailer_send.assert_not_called()
